split index controller
