+++ src/main/java/miu/edu/cs545/controller/IndexController.java
@@ -51,17 +51,10 @@ public class IndexController {
     private final ProductService productService;
     private final SellerService sellerService;
 
-    @Value("${app.default.tax}")
-    private Double taxRate;
-
-    @Value("${app.default.shippingFee}")
-    private Double shippingFee;
-
     @Autowired
     private HomeService homeService;
     @Autowired
     private BuyerService buyerService;
-
 
 
     @Autowired
@@ -70,62 +63,6 @@ public class IndexController {
         this.accountService = accountService;
         this.productService = productService;
         this.sellerService = sellerService;
-    }
-
-    @ModelAttribute("myCart")
-    public Cart getMyCart() {
-        Cart myCart = new Cart();
-        HashMap<String, OnlineOrder> orders = new HashMap<>();
-
-        OnlineOrder order1 = new OnlineOrder();
-        order1.setOrderno("ORD_1#0001");
-        order1.setDateCreate(new Date());
-        order1.setDateShipping(new Date());
-        order1.setShippingFee(16.00);
-        order1.setTax((84 + 150) * 0.08);
-        order1.setTotal(84 + 150 + ((84 + 150) * 0.08) + 16.00);
-
-        Product product = new Product();
-        product.setName("Beige knitted elastic runner shoes");
-        product.setPrice(84.00);
-        product.setPhoto("assets/images/products/table/product-1.jpg");
-        product.setPoint(10);
-
-        Product product2 = new Product();
-        product2.setName("Beige knitted elastic runner shoes");
-        product2.setPrice(75.00);
-        product2.setPhoto("assets/images/products/table/product-2.jpg");
-        product2.setPoint(5);
-
-        List<OrderDetail> detailList = new ArrayList<>();
-        OrderDetail detail1 = new OrderDetail();
-        detail1.setProduct(product);
-        detail1.setSellPrice(product.getPrice());
-        detail1.setQty(2);
-
-        OrderDetail detail2 = new OrderDetail();
-        detail2.setProduct(product2);
-        detail2.setSellPrice(product2.getPrice());
-        detail2.setQty(1);
-
-        detailList.add(detail1);
-        detailList.add(detail2);
-
-        order1.setOrderDetailList(detailList);
-
-        myCart.setOrderList(orders);
-
-        Seller seller = new Seller();
-        seller.setUsername("microsoft");
-        seller.setFirstName("Microsoft");
-        seller.setLastName("Inc");
-        product.setSeller(seller);
-        product2.setSeller(seller);
-
-        orders.put(seller.getUsername(), order1);
-        myCart.setOrderList(orders);
-
-        return myCart;
     }
 
     @GetMapping("/")
@@ -181,20 +118,9 @@ public class IndexController {
         return "buyer/user";
     }
 
-    @GetMapping("/checkout")
-    @PreAuthorize("hasRole('ROLE_BUYER')")
-    public String showCheckout() {
-        return "buyer/checkout";
-    }
-
     @GetMapping("/access-denied")
     public String denyAccess() {
         return "/access-denied";
-    }
-
-    @GetMapping("/login")
-    public String showLoginForm() {
-        return "/buyer/login";
     }
 
     @GetMapping("/about")
@@ -207,82 +133,4 @@ public class IndexController {
         return "buyer/contact";
     }
 
-    @GetMapping("/shopping-cart")
-    public String showCart(Model model) {
-        Cart myCart = getMyCart();
-        model.addAttribute("myCart", myCart);
-        return "/buyer/shopping-cart";
-    }
-
-    @PostMapping("/add-to-cart/{productId}")
-    public @ResponseBody
-    Integer addToCart(@PathVariable(name = "productId") Integer productId, Model model) {
-        Cart cart = (Cart) model.asMap().get("myCart");
-        HashMap<String, OnlineOrder> orders = cart.getOrderList();
-
-        OnlineOrder order = null;
-        Product product = null;
-        OrderDetail orderDetail = null;
-
-        Boolean alreadyInCart = false;
-
-        for (Map.Entry item : orders.entrySet()) {
-            if (alreadyInCart) {
-                break;
-            }
-            order = (OnlineOrder) item.getValue();
-            for (OrderDetail detail : order.getOrderDetailList()) {
-                Product temp = detail.getProduct();
-                if (productId == temp.getId()) {
-                    product = temp;
-                    orderDetail = detail;
-                    alreadyInCart = true;
-                    break;
-                }
-            }
-        }
-
-        if (!alreadyInCart) {
-            Optional<Product> opt = productService.getById(productId);
-            if (opt.isPresent()) {
-                product = opt.get();
-                String sellerId = product.getSeller().getUsername();
-                orderDetail = new OrderDetail();
-
-                if (cart.getOrderList().containsKey(sellerId)) {
-                    order = cart.getOrderList().get(sellerId);
-                } else {
-                    order = new OnlineOrder();
-                    orderDetail = new OrderDetail();
-                    order.setTax(0.00);
-                    order.setShippingFee(shippingFee);
-                    order.setTotal(shippingFee);
-
-                    orderDetail.setProduct(product);
-                    orderDetail.setQty(0);
-                    orderDetail.setSellPrice(product.getPrice());
-
-                    List<OrderDetail> detailList = new ArrayList<>();
-                    detailList.add(orderDetail);
-                    order.setOrderDetailList(detailList);
-                }
-            } else {
-                return 0;
-            }
-        }
-
-        Double newTax = order.getTax() + product.getPrice() * taxRate;
-        Double newTotal = order.getTotal() + product.getPrice() + product.getPrice() * taxRate;
-        order.setTax(newTax);
-        order.setTotal(newTotal);
-
-        Integer newQty = orderDetail.getQty() + 1;
-        orderDetail.setQty(newQty);
-
-        if (!alreadyInCart) {
-            cart.getOrderList().put(product.getSeller().getUsername(), order);
-        }
-
-        return 1;
-    }
 }
